Added Parse Server hosting providers
Parse Server hosting providers are popping up and I believe that this is what lot of us are looking for. I believe we should support these new services by placing this section on top of the page - because they will probably need to gain attention to be able to provide their services for us.